refactor(tests): remove deprecated __tests__ folder for libs
