+++ src/libs/algolia/analytics/logClickOnOffer.test.ts
