+++ src/libs/firebase/analytics/useInit.native.test.ts
