+++ src/libs/firebase/analytics/utils.test.ts
