+++ src/libs/network/OfflinePage.native.test.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 
 import { render, screen } from 'tests/utils'
 
-import { OfflinePage } from '../OfflinePage'
+import { OfflinePage } from './OfflinePage'
 
 describe('<OfflinePage />', () => {
   it('should match snapshot with default message', async () => {
