+++ src/libs/network/useNetInfo.native.test.ts
